services/py/notification/app/routes/notifications.py
from typing import Annotated
from uuid import UUID

import jwt
from fastapi import APIRouter, Depends, Header, Query

from common.errors import AppError
from app.domain.notification import (
    NotificationCreate,
    NotificationResponse,
    NotificationListResponse,
)
from app.services.notification_service import NotificationService

router = APIRouter(prefix="/notifications", tags=["notifications"])


def _get_notification_service() -> NotificationService:
    from app.main import get_notification_service
    return get_notification_service()


def _get_current_user_claims(authorization: Annotated[str, Header()]) -> dict:
    from app.main import app_settings

    if not authorization.startswith("Bearer "):
        raise AppError("Invalid authorization header", status_code=401)
    token = authorization[7:]
    try:
        payload = jwt.decode(
            token, app_settings.jwt_secret, algorithms=[app_settings.jwt_algorithm]
        )
        return {
            "user_id": UUID(payload["sub"]),
            "role": payload.get("role", "student"),
            "is_verified": payload.get("is_verified", False),
        }
    except (jwt.PyJWTError, ValueError, KeyError) as exc:
        raise AppError("Invalid token", status_code=401) from exc


def _to_response(n: "Notification") -> NotificationResponse:
    from app.domain.notification import Notification
    return NotificationResponse(
        id=n.id,
        user_id=n.user_id,
        type=n.type,
        title=n.title,
        body=n.body,
        is_read=n.is_read,
        created_at=n.created_at,
    )


@router.post("", response_model=NotificationResponse, status_code=201)
async def create_notification(
    body: NotificationCreate,
    claims: Annotated[dict, Depends(_get_current_user_claims)],
    service: Annotated[NotificationService, Depends(_get_notification_service)],
) -> NotificationResponse:
    notification = await service.create(
        user_id=claims["user_id"],
        type=body.type,
        title=body.title,
        body=body.body,
    )
    return _to_response(notification)


@router.get("/me", response_model=NotificationListResponse)
async def list_my_notifications(
    claims: Annotated[dict, Depends(_get_current_user_claims)],
    service: Annotated[NotificationService, Depends(_get_notification_service)],
    limit: Annotated[int, Query(ge=1, le=100)] = 20,
    offset: Annotated[int, Query(ge=0)] = 0,
) -> NotificationListResponse:
    items, total = await service.list_my(claims["user_id"], limit, offset)
    return NotificationListResponse(
        items=[_to_response(n) for n in items],
        total=total,
    )


@router.patch("/{notification_id}/read", response_model=NotificationResponse)
async def mark_as_read(
    notification_id: UUID,
    claims: Annotated[dict, Depends(_get_current_user_claims)],
    service: Annotated[NotificationService, Depends(_get_notification_service)],
) -> NotificationResponse:
    notification = await service.mark_as_read(notification_id, claims["user_id"])
    return _to_response(notification)
